Determine evaluation times of different algorithms in Application
Instead of applying all methods to one entry in the tree and then
moving to the next entry, the code now processes the whole tree using
one method, and then the whole tree again using the next method, etc.
This is slower, but it allows me to compare the evaluation times of
different MVA algorithms. (Might need some refactoring later.)